add CHeckHelper
